requirements.txt
asgiref==3.9.2
Django==5.0.14
mysqlclient==2.2.7
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.15.2
django-modeltranslation==0.18.11
django-ratelimit==4.1.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
requests==2.32.3
django-filter==23.5
django-autoslug==1.9.9
